--- a/python/chronos/src/bigdl/chronos/forecaster/base_forecaster.py
+++ b/python/chronos/src/bigdl/chronos/forecaster/base_forecaster.py
@@ -34,6 +34,7 @@ from bigdl.nano.utils.common import invalidOperationError, invalidInputError
 from bigdl.chronos.data.tsdataset import TSDataset
 from bigdl.chronos.pytorch.context_manager import DummyForecasterContextManager,\
     ForecasterContextManager
+from bigdl.chronos.pytorch.utils import _pytorch_fashion_inference
 
 
 class BasePytorchForecaster(Forecaster):
@@ -350,7 +351,6 @@ class BasePytorchForecaster(Forecaster):
             num_nodes = 1 if sc.get('spark.master').startswith('local') \
                 else int(sc.get('spark.executor.instances'))
             if batch_size % self.workers_per_node != 0:
-                from bigdl.nano.utils.common import invalidInputError
                 invalidInputError(False,
                                   "Please make sure that batch_size can be divisible by "
                                   "the product of worker_per_node and num_nodes, "
@@ -362,7 +362,6 @@ class BasePytorchForecaster(Forecaster):
                                      batch_size=batch_size)
         else:
             from bigdl.chronos.pytorch import TSTrainer as Trainer
-            from bigdl.nano.utils.common import invalidInputError
 
             # numpy data shape checking
             if isinstance(data, tuple):
@@ -702,9 +701,6 @@ class BasePytorchForecaster(Forecaster):
                  where result is a numpy array with shape (num_samples, horizon, target_dim)
                  if data is a xshard item.
         """
-        from bigdl.chronos.pytorch.utils import _pytorch_fashion_inference
-        from bigdl.nano.utils.common import invalidInputError
-
         if quantize or acceleration:
             self.thread_num = set_pytorch_thread(self.optimized_model_thread_num, self.thread_num)
 
@@ -806,8 +802,6 @@ class BasePytorchForecaster(Forecaster):
 
         :return: A numpy array with shape (num_samples, horizon, target_dim).
         """
-        from bigdl.chronos.pytorch.utils import _pytorch_fashion_inference
-        from bigdl.nano.utils.common import invalidInputError
         if self.distributed:
             invalidInputError(False,
                               "ONNX inference has not been supported for distributed "
@@ -885,9 +879,6 @@ class BasePytorchForecaster(Forecaster):
 
         :return: A numpy array with shape (num_samples, horizon, target_dim).
         """
-        from bigdl.chronos.pytorch.utils import _pytorch_fashion_inference
-        from bigdl.nano.utils.common import invalidInputError
-
         if self.distributed:
             invalidInputError(False,
                               "Openvino inference has not been supported for distributed "
@@ -966,9 +957,6 @@ class BasePytorchForecaster(Forecaster):
 
         :return: A numpy array with shape (num_samples, horizon, target_dim).
         """
-        from bigdl.chronos.pytorch.utils import _pytorch_fashion_inference
-        from bigdl.nano.utils.common import invalidInputError
-
         if self.distributed:
             invalidInputError(False,
                               "Jit inference has not been supported for distributed "
@@ -1070,9 +1058,6 @@ class BasePytorchForecaster(Forecaster):
 
         :return: A list of evaluation results. Each item represents a metric.
         """
-        from bigdl.chronos.pytorch.utils import _pytorch_fashion_inference
-        from bigdl.nano.utils.common import invalidInputError
-
         # data transform
         if isinstance(data, TSDataset):
             _rolled = data.numpy_x is None
@@ -1187,8 +1172,6 @@ class BasePytorchForecaster(Forecaster):
 
         :return: A list of evaluation results. Each item represents a metric.
         """
-        from bigdl.chronos.pytorch.utils import _pytorch_fashion_inference
-        from bigdl.nano.utils.common import invalidInputError
         if self.distributed:
             invalidInputError(False,
                               "ONNX inference has not been supported for distributed "
@@ -1298,8 +1281,6 @@ class BasePytorchForecaster(Forecaster):
                  with shape (num_samples, horizon, target_dim)
 
         """
-        from bigdl.chronos.pytorch.utils import _pytorch_fashion_inference
-
         if self.distributed:
             invalidInputError(False,
                               "predict interval has not been supported for distributed "
@@ -1394,7 +1375,6 @@ class BasePytorchForecaster(Forecaster):
             self.internal.save(checkpoint_file)
         else:
             if not self.fitted:
-                from bigdl.nano.utils.common import invalidInputError
                 invalidInputError(False,
                                   "You must call fit or restore first before calling save!")
             # user may never call the fit before
@@ -1465,7 +1445,6 @@ class BasePytorchForecaster(Forecaster):
         :return: a forecaster instance.
         """
         from bigdl.chronos.pytorch import TSTrainer as Trainer
-        from bigdl.nano.utils.common import invalidInputError
         # TODO: optimizer is refreshed, which is not reasonable
         if not self.distributed:
             invalidInputError(False, "The forecaster has become local.")
@@ -1533,7 +1512,6 @@ class BasePytorchForecaster(Forecaster):
         '''
         import onnxruntime
         from bigdl.chronos.pytorch import TSInferenceOptimizer as InferenceOptimizer
-        from bigdl.nano.utils.common import invalidInputError
         if sess_options is not None and not isinstance(sess_options, onnxruntime.SessionOptions):
             invalidInputError(False,
                               "sess_options should be an onnxruntime.SessionOptions instance"
@@ -1580,7 +1558,6 @@ class BasePytorchForecaster(Forecaster):
                `OMP_NUM_THREADS` is suggested to be same as `thread_num`.
         '''
         from bigdl.chronos.pytorch import TSInferenceOptimizer as InferenceOptimizer
-        from bigdl.nano.utils.common import invalidInputError
 
         if self.distributed:
             invalidInputError(False,
@@ -1623,7 +1600,6 @@ class BasePytorchForecaster(Forecaster):
                `OMP_NUM_THREADS` is suggested to be same as `thread_num`.
          '''
         from bigdl.nano.pytorch import InferenceOptimizer
-        from bigdl.nano.utils.common import invalidInputError
 
         if self.distributed:
             invalidInputError(False,
@@ -1656,7 +1632,6 @@ class BasePytorchForecaster(Forecaster):
         :param quantized_dirname: The dir location you want to save the quantized onnx file.
         """
         from bigdl.chronos.pytorch import TSInferenceOptimizer as InferenceOptimizer
-        from bigdl.nano.utils.common import invalidInputError
         if self.distributed:
             invalidInputError(False,
                               "export_onnx_file has not been supported for distributed "
@@ -1682,7 +1657,6 @@ class BasePytorchForecaster(Forecaster):
         :param quantized_dirname: The dir location you want to save the quantized openvino file.
         """
         from bigdl.chronos.pytorch import TSInferenceOptimizer as InferenceOptimizer
-        from bigdl.nano.utils.common import invalidInputError
         if self.distributed:
             invalidInputError(False,
                               "export_openvino_file has not been supported for distributed "
@@ -1775,7 +1749,6 @@ class BasePytorchForecaster(Forecaster):
                the value can be arbitrary.
         """
         from bigdl.nano.pytorch import InferenceOptimizer
-        from bigdl.nano.utils.common import invalidInputError
         from pathlib import Path
         if self.distributed:
             invalidInputError(False,
@@ -1886,7 +1859,6 @@ class BasePytorchForecaster(Forecaster):
                default where no limit is set
         """
         # check model support for quantization
-        from bigdl.nano.utils.common import invalidInputError
         from bigdl.chronos.pytorch import TSInferenceOptimizer as InferenceOptimizer
         if not self.quantize_available:
             invalidInputError(False,
@@ -2003,7 +1975,6 @@ class BasePytorchForecaster(Forecaster):
 
         :return: A Forecaster Model.
         """
-        from bigdl.nano.utils.common import invalidInputError
         invalidInputError(isinstance(tsdataset, TSDataset),
                           f"We only supports input a TSDataset, but get{type(tsdataset)}.")